Restructuring of the classes. The Math.body file which contains the MathRender class was split in the following way Math.base contains the base class with the database related stuff and provides an abstract interface Math.source and Math.MathJax handle the plain tex string output. There are two classes since they will differntiate in the future I think. Math.texvc contains the "old" implementation of png generation with all the file handling related stuff Math.LaTeXML is a new (experimental) way which uses the Webservice LaTeXML to convert LaTeX to XML. LaTeXML can use MathJax to display MathML if the client browser does not support it. Other implementation of math renderer can be added in the same style. Change-Id: I193a4985b145a70e849a6d45c57c757e98db3add
Added MathML rendering by MathJax for Browsers other than firefox

Change-Id: I2e6d4c2dea3ce612d1dbcc463bda624001c5fd55

added customized file for mathml to mathjax conversion by mathjax

Change-Id: I88d5bc8b9bc6872018b96283ab5019c47ebf6fc2

Styling of the files with stylize.php

Change-Id: Ibb6032de30efd5c6b0d55b798bc20ed2c710e154

fixed some issues according to the comments of Nikerabbit

Change-Id: I98001e4cefe5fffb6f9edfa2830576c99c746874

Renaming of the files according to suggestion of Nikerabbit

Change-Id: I4c04f8dc54b1fe26965f47c3fb49a08629105c69

Restructuring of the classes.

The Math.body file which contains the MathRender class was split in the following way
Math.base contains the base class with the database related stuff and provides an abstract interface
Math.source and Math.MathJax handle the plain tex string output. There are two classes
since they will differntiate in the future I think.
Math.texvc contains the "old" implementation of png generation with all the file handling related stuff
Math.LaTeXML is a new (experimental) way which uses the Webservice LaTeXML to convert LaTeX to XML.
LaTeXML can use MathJax to display MathML if the client browser does not support it.
Other implementation of math renderer can be added in the same style.

Improved version!

Change-Id: I9b1d68c4faa8d177d8d0088fa1a5879caed4f1fe