Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

(1 inline comment)

I have tested it. For MathJaX it works. Even for MathML is works fine (as before) if one changes the parameter.

PS: I'd be interested if there is an easy way to decide to use MathJax or not dependent on the users browser.

Patch-set: 3