Update patch set 4
Patch Set 4:

@The DJ

What I wanted to see is the following.
-A major commit message with the main idea and backgound of this modification.
-Comments for the minor changes that relat to the changesets.

I still didn't figure out how to do that.

Furthermore I'm looking for the right place to disuss futher modification of the extension. For the core component there is a roadmap page at the wiki. I created a similar page for the math extension at:
http://www.mediawiki.org/wiki/Extension:Math/Roadmap
but I question if anyone will find that page.

Patch-set: 4