Update patch set 1
Patch Set 1:

Hi TheDJ,
unfortunately I don't get it.
What is the general stragegy concerning displaystyle?
How do you detect that the displaystyle is wanted or not?
What does this mode stand for and how is it set?
How would the user define that he want's to get a displaystyle equation?

physikerwelt

Patch-set: 1