Update patch set 19
Patch Set 19: (22 inline comments)

I'm helping out a bit with the coding conventions and documentation.  I've rebased this and addressed the convention/doc items I mentioned earlier.

I tested this with both the PNG and source versions.  It seemed to work fine once I got around a file/directory creation permission issue.  I'm pretty sure that is just because it's running in a vagrant (and I didn't try to configure that before).

However, it does still need a bit more testing.

Patch-set: 19