Update patch set 20
Patch Set 20:

There are no unit tests for the math extension yet. So I rendered the WP:MATH page for testing with both version in the Chrome browser and ran a diff afterwards. The result looked fine for me. To do so you have to delete the generated images and to set the check math files option to true.

Patch-set: 20