Fix Math for databases that use Blob classes.
Some databases like sqlite use a Blob class as the output of encodeBlob
which is used by getInputHash. This results in a fatal error being thrown
when writeDBEntry constructs the wfDebug line since the Blob class cannot
be converted into a string.

Fix the issue by checking to see if the return from getInputHash is a Blob
class and if so use the blob's fetch method.

Change-Id: I3fb670aca4444f149e4bc2bd4872479b0da8e8fb