Update patch set 4
Patch Set 4: Code-Review-1

(7 comments)

These tests are OK, but I think they could be improved by stubbing out the database dependency using test doubles.

http://www.phpunit.de/manual/current/en/test-doubles.html#test-doubles.mock-objects

What I would do is this:
* Move as much logic as possible out of the writeDBentry and readFromDB. You already have $this->getInputHash(), which is good -- couldn't you also have $this->getOutputHash()? (Alternately, a getHash() method that could serve both).
* Write tests for those functions.
* Leave the leftover database methods untested. Your assumption should be that any failure of those functions points to a failure outside the extension, and you shouldn't be testing for those. If you're not convinced that failures are necessarily the fault of something outside the extension, go back to step 1, because it means that some logic could still be moved away from these methods.
* Test functions that call database methods by replacing the database methods with mocks. The idea is that rather than asserting that some value equals another value, you assert that a function was called n times with parameters x, y, z. (This allows you to test your extension in isolation. It also makes the test suite faster, because it removes the dependency on external resources.) See the section on mock objects in the phpunit manual: http://www.phpunit.de/manual/current/en/test-doubles.html#test-doubles.mock-objects 

You can see some examples here: https://gerrit.wikimedia.org/r/gitweb?p=mediawiki/extensions/EventLogging.git;a=blob;f=tests/RemoteSchemaTest.php
The RemoteSchema class that I am testing in that file fetches content from memcached and via HTTP, but I mock those out.

I will make one last recommendation, but I will preface it with the caveat that t reflects my personal tastes rather than some kind of agreed-upon best practice. The recommendation is this: make everything public. There is very little to be gained from constraining visibility. There is value in indicating to people working with your code what is and is not part of a public API, but you can do that just as well with comments and with things like prefixing the method name with '_'. This makes code much easier to test and to use.

---
I found this conversation illuminating (from http://www.codequarterly.com/2011/rich-hickey/):
Fogus: Following that idea—some people are surprised by the fact that Clojure does not engage in data-hiding encapsulation on its types. Why did you decide to forgo data-hiding?
Hickey: Let’s be clear that Clojure strongly emphasizes programming to abstractions. At some point though, someone is going to need to have access to the data. And if you have a notion of “private”, you need corresponding notions of privilege and trust. And that adds a whole ton of complexity and little value, creates rigidity in a system, and often forces things to live in places they shouldn’t. This is in addition to the other losing that occurs when simple information is put into classes. To the extent the data is immutable, there is little harm that can come of providing access, other than that someone could come to depend upon something that might change. Well, okay, people do that all the time in real life, and when things change, they adapt. And if they are rational, they know when they make a decision based upon something that can change that they might in the future need to adapt. So, it’s a risk management decision, one I think programmers should be free to make.
If people don’t have the sensibilities to desire to program to abstractions and to be wary of marrying implementation details, then they are never going to be good programmers.
---

Patch-set: 4