Update patch set 4
Patch Set 4:

(Speaking of 'writeDBentry' and 'readFromDB' -- is there any reason not to make the names more symmetrical? I'd prefer "writeDatabaseEntry" and "readDatabaseEntry". (I prefer 'Database' to DB because having two capital letters in a row breaks the camelCase flow and is a source of frequent typing errors for me -- e.g. I often mistype "$wgDBname" as "$wgDbName".)

Patch-set: 4