fix MathSource::writeCache
Fix of the logical error that inverts the meaning of recall.
If isRecall() is true, the formula was obtained from the database and needs
not to be stored in the database again. If isRecall() was not true the formula should
be stored in the database to prevent re-rendering on every change
(or action=purge call) of the page.

Change-Id: Ib1f361bdfeba1c4ea790dfd0330adc8d28d849fc