Update patch set 4
Patch Set 4:

(7 comments)

Ok the use of these mock functions is quite convincing.
However I'm not sure if this shrinks the test coverage. i.e. the creation of the math table and the debug fields can hardly be tested without writing to database. Furthermore the encoding problems that occurred due to some non-standard characters in the MathML output would have been undiscovered as well. 

But in the end of the day, these things should be tested in the runtime environment and not in the development phase. Therefore unit tests are not the correct context to test this, because we don't expect the user to run the test in her/his system. Maybe these kind of tests should be put to LoadExtensionSchemaUpdates? Did I get that right?

I’m wondering why there are functions like setupTestDB in the MediaWikiTestCase class? Is that approach compatible with the Mocking approach?
If_ indicates do not use until you know exactly what you are doing, public methods are a good idea. But since people recommended not to use _ names for private methods I switched to protected.

Patch-set: 4