Update patch set 1
Patch Set 1: Code-Review-1

If $wgUseSquid is true and $this->isRecall() is true, it means that the image was stored in the database, but perhaps not in Squid. (It could have been purged, etc.) In those cases, you still want to put it in Squid, probably, since it's much closer to the edge and cheaper to access than the database. But your change would return early and short-circuit Squid caching.

Since you're cleaning things up, I might as well mention that 'isRecall' is not a very descriptive name, and it collapses the distinction between Squid and the database in a way that is not easy to disentangle. If you don't feel like fixing that yourself (or if you do, but not for this changeset), could you please add a note about this to a TODO file in the repository root?

(In fact, it's probably a good idea to start a TODO file either way. I know you have the roadmap on MediaWiki.org, but that tracks higher-level ambitions, rather than various warts in the code that it would be nice to clean up.)

Patch-set: 1