Specialpage MathDebug
This commit introduces the specialpage mathdebug.
This page tests the rendering of typical math-expression using different
rendering engines.
It is accessible if $wgDebugMath is set to true for users with sysadmin rights.
Especially for small wikis the math extension was a crucial point that caused
errors. Now admins can see if math rendering works correctly on their server.

Change-Id: I39f853a5a2a1a5dabdd0350b11b4ed3c0fe17ef8