Update patch set 8
Patch Set 8:

(3 comments)

Thanks for your comments.
The sandbox has another purpose.
This particular set of tests is the most complete list of texvc test I could find. Furthermore it provides an overview how the different renderer behaves.
Maybe it is more useful for developer than for sysadmins, but I think it is helpful for both groups.

Patch-set: 8