Update patch set 9
Patch Set 9: Code-Review-2

I thought about this some more, and I'm not okay with this in the current form.

It would work as a sandbox where you input an arbitrary math expression and render it various ways (again, similar to TemplateSandbox).

The main problem with this is basically testing, but you're making admins do it manually.  We should instead figure out how to automate this using some combination of parser/unit/integration/Cucumber tests, instead of hard-coding it into the main (non-test) code.

Patch-set: 9