Update patch set 4
Patch Set 4:

mh.. after I had problem with the update of git review the file was empty for some reason.
However for the current version, this test does not test to much. But as I have made the experience that it is much easier to find reviewer for minimal changes, I'm trying to break down things to minmal changes now.

Patch-set: 4