debug log was attempting to show a Blob object
A wfDebugLog() was attempting to concatenate a string with a Blob object
which does not support conversion to a string.  Since that was used to
show some hash, display the current string hash instead.
Furthermore the input-hash, which can be calculated from the TeX input
string was displayed, which is not very helpful. The variable hash is 
calculated by texvc and can be used to search for the rendered image in
the filesystem.

Change-Id: I9943fd51d3021bf2d62a29f33de0858803763f86