Update patch set 1
Patch Set 1:

It's somewhat complicated.

Displaystyle (block mode) is desired, however the place that we are adding it in was not ideal. It basically made our 'display style' PART of the tex formula. With the prefilter, we move this addition into the phase before the rendering, but after the definition of the formula. I'll try to amend the commit message a bit.

Patch-set: 1
Label: Verified=0