Update patch set 1
Patch Set 1: Code-Review-1

To my mind MathJax should only render the tags converted by
https://gerrit.wikimedia.org/r/gitweb?p=mediawiki/extensions/Math.git;a=blob;f=MathMathJax.php

maybe we should change 'class' => 'tex', to 'class' => 'MathJax'
in oder to avoid confusion.

By the way one of the recent changes broke my Test for Debugging at 
https://gerrit.wikimedia.org/r/gitweb?p=mediawiki/extensions/MathSearch.git;a=blob;f=SpecialMathDebug.php;h=0e228caaa1d95618209582ce5d23611966fb70b9;hb=refs/heads/master#l39
.
The images is shown for a very short time and than it is replaced by the MathJaX stuff. That is not intended. At least at this place.

Patch-set: 1