Update patch set 1
Patch Set 1:

Ah, I realize that this is not working in HTML/CSS mode due to the default styling that of mtext that we still have in. No I understand why Nageh changed those as well. I'll add the style changes to this changeset as well, which should enable multiple font styles.

Patch-set: 1
Label: Verified=0