Update patch set 22
Patch Set 22: Code-Review-1

(9 comments)

I investigated the tests some.  I think part of it is that we're using the PHP HTTP locally (which may have some timeout bugs), but I think Jenkins is using Vagrant.

However, I'm not positive, and this doesn't explain why it missed the i18n whitespace (which is considered test output).

You may want to put a deliberately failed assertion, to make sure Jenkins is running this file.

I agree we should mock the HTTP.  But if we can't figure out why the PHP HTTP backend isn't working, we might want to put a note strongly recommending the cURL.

Also, am I right that you're testing the timeout with infinite recursion?  That reminds me that there doesn't seem to be any validation of input, which is another reason not to recommend this too heavily yet.

Patch-set: 22
Label: Code-Review=-1