Update patch set 24
Patch Set 24:

> First, let me announce I've released a new web service 
> for LaTeXML that is free of experimental code, with the 
> intention of being robust and installable entirely 
> through a linux package manager. It's based on 
> PSGI/Plack, written in their middleware language. Here is > the announcement: http://lists.jacobs-university.de/pipermail/project-latexml/2013-May/001336.html

Interesting.  It should be compatible with this code, if I understand "HTTP POST/GET requests with the familiar syntax" correctly.  You may want to use the bug (https://bugzilla.wikimedia.org/show_bug.cgi?id=43222) to discuss broader points that aren't specific to this code review.

> Well, TeX is a turing complete language, and LaTeXML 
> reimplements the entire TeX interpreter, so building a 
> black-list won't be easy - you probably need to outlaw
> over 50 macros and you're bound to miss some attack angle.

Yes, I'm aware of that.  Of course, a whitelist must be used, not a blacklist.  texvc (the PNG backend used by default) already takes this approach.

> If you want to have a narrow list of whitelisted macros, 
> you're probably better of using MathJaX which can parse a 
> very minimal subset of TeX.

We want to support both, at least for now.

> Instead, LaTeXML does a meta-restriction - it blocks all 
> IO in web service mode, and accepts a "timeout" option, 
> which would kill the processing job after N seconds.

Is this available for the web service we're currently using, or just your new one?  As you can see, we're currently doing a client-side timeout, though there see to be issues when not using the cURL HTTP client.

> For Wikipedia formulas, I assume a timeout of 10 or 15 
> seconds would be sufficient, though Moritz would have to 
> say what his data suggests.

It's currently using a rather high timeout of 240, due to an outlier article in English Wikipedia that takes a long time.  Individual users of the extension can customize this, though.

Patch-set: 24
Label: Verified=0