Update patch set 5
Patch Set 5:

Okay, I'm fine with setting them to none.

However, I tested this, and it doesn't work since the extension list is getting clobbered.

To see this:

1. Temporarily comment out the href line in texvc.js (I was having a bit of trouble testing the other features of safe mode).
2. Try to make a href in a MathJax formula.

The link will work.  To fix safe mode, remove the new extension list in ext.math.mathjax.enabler.js, and put Safe.js in the list in config/TeX-AMS-texvc_HTML.js.

It's somewhat confusing that there are two separate configs, and the arrays aren't merged (there's a comment in upstream MathJax wondering if arrays should be concatenated).

Patch-set: 5