Enabling MathJaX in LaTeXML rendering mode exept for Firefox
Due to a bad mathml support for all browsers expect Firefox,
the LaTeXML rendering mode will use MathJaX's MML to something
that can be displyed by the brower rendering mode.
This is not the same as enabling MathJaX since it it is much faster
since the MathML must only be transformed to e.g. SVG. In contrast
to that MathJaX must compile the tex code in oder to get the SVG
output at the client side which takes much longer.

Change-Id: I188d85a914483c5c1ce5a16debfece93a3c438a7