Enabling MathJaX in LaTeXML rendering mode except for Firefox
Due to a bad MathML support for all browsers, except Firefox,
the LaTeXML rendering mode will use MathJaX's MML to something
that can be displayed by the browser rendering mode.
This is not the same as enabling MathJaX since it is much faster
since the MathML must only be transformed to e.g. SVG. In contrast
to that MathJaX must compile the TeX code in order to get the SVG
output at the client side which takes much longer.

PS: Move functionality to JavaScript to avoid caching problems.

Change-Id: I6ad17a04e9d4afff71ebbb3f4bc3cb6d4fd15e5d