Update patch set 1
Patch Set 1:

I'm not sure if we really want this.
For me it sounds more like a feature for power users. Those might be able to change that via a user script anyhow.
So this is a little bit contradictory to the intention to use MathML / MathJaX as the standard output format.

Patch-set: 1