Update patch set 3
Patch Set 3:

Hi,

I agree on index of but is has to be
 if (navigator.userAgent.indexOf("Firefox") == -1){
Since only something should be done if the browser is not! Firefox, since Firefox is the only browser that can handle mathml well.
The dependency does not help. Since if the MathML base module is loaded it renders all equations. (If there are MathML equation the tex code is extracted and they are rendered as TeX.)
This could be changed by changing the class of the Math output element from 'tex' to something else.
However then, we have to teach MathJaX to listen for another tag if configured.
The bad thing with this approach is, that Firefox had to load MathJaX even though it does never use MathJaX.
Is there no way to model the dependency inside JS and not via PHP?

Patch-set: 3
Label: Verified=0