Update patch set 12
Patch Set 12:

Don't add the Enabler as a global variable. Make it a local variable or add it to the mw.libs host object.

The "( function ( mw, $ ) {" closure should be the outer most thing in the file, so move the function inside of that and then at the end expose it through mw.libs.SomethingEnabler = SomethingEnabler;

Patch-set: 12
Label: Verified=0