Update patch set 17
Patch Set 17: Code-Review-1

It looks like it is currently doing this:

* base: sets up abstract configuration
* enablers: call base with one of 2 javascript file names
* MathJax lib (third-party): uses the config and loads the file from the extension directory raw and unminified.

Ideally we don't have the weird abstraction of loading config through plain text in a non-standard eval() <script> tag using a global static variable mathJax.config that is then sometimes re-used and extended with a different configuration when the next one is needed. I don't know why it preserved (part?) of the old values as they seem no longer relevant the next time.

I know this is somewhat unfair to Physikerwelt, but it has to be said: MathJax is a piece of shit javascript-wise and loading it as a whole hurts but works without doing any crazy stunts. But loading parts of it gets really tricky especially since MathJax seems to require this kind of structure. However the way this is currently done is imho simply unacceptable and it has to change. I can't do this myself right now due to other obligations, but I can give advice if needed.

Patch-set: 17
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1