Update patch set 17
Patch Set 17:

There are definitely parts of the MathJax that are confusing, indirect, and don't meet MW standards.  In some cases, it also just clashes (our ResourceLoader vs. their loader).

That said, the MathJax team is working with us (albeit mainly on i18n so far), so please be civil.  I'd like to be able to add them as reviewers on this kind of change.

MathJax obviously allows loading config files as extension, but I don't think that's the only option.  It should be possible to call MathJax.Hub.Config( mathJax.config ) directly in the enabler "subclasses".

However, one problem is that while the TeX-AMS-texvc_HTML.js code is ours (so we could restructure it pretty easily), MML_HTMLorMML-full.js is not.

If we modified how MML_HTMLorMML-full was loaded, it would require extra work every time we synced with upstream.  
Also, MML_HTMLorMML-full.js is pretty complex code.  It's really not an "extension", but a particular build of some important code.

The text/x-mathjax-config thing is weird, but is explained at http://docs.mathjax.org/en/v2.2-latest/configuration.html#using-in-line-configuration-options .  If we don't want to use it, we can configure delayStartupUntil=configured .  The question then is *how* to do that single configuration.  It can be done in the URL, but that's not workable if we want to use RL.  So the options I see are:

1. (Maybe, not tested). Use text/x-mathjax-config just to say that it's delayed startup.
2. Do a local mini-fork (we already have TeX-AMS-texvc_HTML.js, so I don't think this is a huge deal) to set delayStartupUntil=configured.

Krinkle, given that, please try to propose what you consider an acceptable approach.

Patch-set: 17
Reviewer: Mattflaschen <519@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0