Update patch set 17
Patch Set 17:

how to go on with this issue.
I think we should improve the structure of the code by means of ours and theirs.
I want to use MathJax as a kind of patch for bad browsers.
Therefore all that should be done is to enable MathJaX and
use it to render MathML as defined by W3C.
I can understand that it makes sense to make the language
of mediawiki the default language of the mathjax context menu 
and some special styling issues... and maybe some other aspects
that are specific for Wikipedia.
However, in general I recommend to use MathJax as it is, and to 
report bugs to the MathJax team, as most of them won't be
Wikipedia specific.
From that point of view, it might be better to include
Math/modules/MathJax
as git submodule and to set up another git repository
that basically mirrors the current stable mathjax version.
The customization done in the config folder i.e. in 
TeX-AMS-texvc_HTML.js should be tried to be put out of
the MathJax folder to my mind.
But, is that still in the scope of this change?

Patch-set: 17