Update patch set 17
Patch Set 17:

Maybe we could formulate a task... so that I know what I have
to get this merged... MathJax with LaTeXML would be really
cool and applicable for a wide audience.
I propose the following:
I write an MathJaX enabler that this independent of
the current one... so this would influence the people
using latexml only (that are just a few at the moment) and have no
influence on the target group that use mathjax to render
texvc with their browsers.

Patch-set: 17
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0