Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Please also update the README.  You can say that the latest stable version (1.21.1) is supported (maybe it goes back farther, but you'll need to verify), but that you need 1.22wmf7 or newer for full functionality.

The test thing is only tangentially related (you're not testing the is_callable logic).  It's alright if you want to put it in this commit, but update the commit message.

Patch-set: 1
Label: Code-Review=-1