Update patch set 1
Patch Set 1:

The default is:

availableFonts: ["STIX","TeX"]
preferredFont: "TeX"
webFont: "TeX"

So the choice will be:
- STIX and TeX installed => local TeX fonts
- only STIX installed => local STIX fonts
- only TeX installed =>  local TeX fonts
- nothing installed => TeX web fonts

The current WikiMedia setup has availableFonts: ['TeX'] so the difference with the default is that in the "only STIX installed" case, TeX Web fonts will be used.

The patch uses preferredFont: ['STIX'] instead, so the difference with the default is that in the "STIX and TeX installed" case, local STIX fonts will be used.

Basically some people prefer TeX fonts because they have the aspect of the Computer Modern fonts used in LaTeX while other prefer STIX fonts because they have a much better Unicode coverage. The STIX fonts are also more common on user systems and so that speeds up the rendering compared to Web fonts.

Perhaps an option would be to keep the current set up but to enable showFontMenu so that people can choose the fonts.

http://docs.mathjax.org/en/latest/options/MathMenu.html

(note: we expect MathJax v2.3 to support more fonts)

Patch-set: 1
Label: Verified=0