Update patch set 2
Patch Set 2:

The "build failure" is:
14:27:44 modules/ext.math.mathjax.enabler.mml.js: line 5, col 41, Strings must use singlequote.
I think this is easy to fix.
The overall status is the following.
The LaTeXML branch has a copy of the normal texvc MathML loader (ext.math.mathjax.enabler.mml) which I modified just a little bit. In the way that MathML is used rather than texvc.
I think it's better to work on
https://gerrit.wikimedia.org/r/#/c/69298/

in general I think we should submit patches to the main branch and not LaTeXML. If one of us does the code review it is OK. Maybe you can have a detailed look at
https://gerrit.wikimedia.org/r/#/c/69298/
and see how the mathjax loader can be improved in a way that it sticks to mediawiki javascript coding guides 
http://www.mediawiki.org/wiki/Manual:Coding_conventions/JavaScript
and is optimal for mathjax as well. 
I granted you access to http://math-test2.instance-proxy.wmflabs.org/wiki/Main_Page
via wikitech
https://wikitech.wikimedia.org/wiki/Main_Page
you can try to login. We can change to the main branch there.

Patch-set: 2
Label: Verified=0