Update patch set 1
Patch Set 1:

I think delayStartupUntil=configured is supposed to be passed as a query string for MathJax.js. Unfortunately, it seems that it is not possible to pass such a parameter with mw.loader.using ; mw.loader.load does not seem to provide a callback and otherwise I would have to use jquery.getScript. So for now I've cheated a bit an modified MathJax.js to set delayStartupUntil to configured. Probably that's something that should be modified upstream.

I've used MathJax's unpacked version to let MediaWiki's loader compress the data. I'm still thinking about the best way to replace MathJax's loader by MediaWiki's loader. For the moment, this branch demonstrates how to do that for the localization data.

Patch-set: 1