Update patch set 2
Patch Set 2:

Yes, I tested it on my local machine and it works very well. I didn't check whether this interferes with LaTeXML, though (note: at the moment MathJax won't process the <math> tag since I removed the preprocessors). This changes the output of the formula so you'll have to edit a wiki page to take the modification into account and not use the cached version.

Do you have texvc installed and configured (note: I updated https://www.mediawiki.org/wiki/Extension:Math)? It seems that texvc is supposed to always be enabled ($wgUseTeX no longer exists) so in theory you should have it installed and working. Perhaps this can be changed, so that the MathJax mode could produce the PNG image only optionally.

Ideally, I think texvc should be dropped in the future, but currently there are not any satisfying alternatives to replace it (from the point of view of users and MediaWiki folks I mean ; personally I'm happy with MathML rendering in Gecko). Hence Peter's idea to keep the images but silently replacing them by MathJax's output would give something close to the current setup (to which people are used), while still bringing improvements. The LaTeXML mode can show the ideal long term alternative: server-side TeX-to-MathML conversion and, for Gecko-based browsers, native MathML rendering.

The advantage to have both the preview (PNG for MathJax or MathML for LaTeXML) and MathJax's <script> tag is that it will work when users have Javascript disabled and it will allow to decide client-side whether to enable MathJax or not. In particular, it seems that if we want to enable the options for anonymous users, the texvc and MathJax modes will eventually have to produce exactly the same HTML output (see bug 48036 comment 4 and later). Unfortunately, this means that the LaTeXML mode won't be available to anonymous users, unless we also merge the output of the three modes (MathML, PNG and TeX source).

Anyway, the exact organization of modes can be decided later. My main concern is now to figure out the best way to load MathJax with MediaWiki's resource loader. I'll look into this later when I'm back to work.

PS: what I do noticed is that you'll have to clear the cookie for the MathJax menu otherwise you get a Javascript error. This is probably due to how I load MathJax with MediaWiki's resource loader, so I'll have to check.

Patch-set: 2