Update patch set 2
Patch Set 2:

..mh
maybe it's to much for one commit
I would not like to require texvc for mathjax... that is somehow contraproductive to my mind

are you familar with grammars like antlr? I would like to rewrite texvc in such a grammar that can be than compiled to php

At least for me it did not work... did you include all files you changed in your commit?

Patch-set: 2