Update patch set 2
Patch Set 2:

I don't understand your problem. It seems that you haven't been able to make texvc work on your system and thus you are asking to not make MathJax mode depends on it. Here I'm just adopting a pragmatic point of view: if we want anonymous to have access to the MathJax mode available then the PNG and MathJax mode should produce exactly the same output (bug 48036). So that would actually mean a PNG+MathJax mode. Of course, it's easy to just create a new MathJax+PNG mode different from the MathJax only mode, so let's do that. Again, this was an experimental commit to test the water and find potential issues, not the definitive version. As I see now:

- the wiki2jax preprocessor will unfortunately still be needed because the PHP code seems to always escape HTML entities even inside <script> and because it does not know the display mode. When this is fixed in MediaWiki, the wiki2jax preprocessor could be dropped and that will speed up a bit MathJax.

- A new MathJax+PNG mode should be created instead. By the previous point, it would need the same output as the PNG mode that is <img src="TeX Source"> rather than the "<img><script>TeX Source</script>" which is what the wiki2jax processor will have to produce. We'll keep a "MathJax only" mode which will be lighter than MathJax+PNG but still slower and heavier than LaTeXML+Firefox.

- The x-mathjax-config (and its eval call) will still be needed unless we modify MathJax.js to hardcode delayStartupUntil to Configured. (I'll try to discuss with MathJax devs about how to modify MathJax.js upstream to take the config into account without requiring an x-mathjax-config script)

- The update to MathJax v2.3 was essentially to test the localization data and the parsing bug with mixed MathML and TeX. This can be wait for when it is released.

- The point of using the unpacked version was to test how to let MediaWiki's resource loader minifies the JS source. This is what requires all the file addition/removal. I guess we can still keep on the already-packed-version, but that seems better to keep MathJax's unpacked version for debugging.

Later, I'll concentrate on continuing the work for migrating MathJax's loader to MediaWiki's loader.

Regarding replacing texvc by a TeX-to-PNG PHP program, I'm personally not interested to work on this. My goal is to help migrating to a MathML-based system (like MathJax or LaTeXML) which is what the MathJax project is pushing for. If in the future I had to work on another server-side converter, it would probably be MathJax-based (like TeX-to-MathML + MathML-to-SVG).

Patch-set: 2