Update patch set 4
Patch Set 4:

So I've updated the patch and only kept the upgrade to MathJax 2.3 (including the new upstream fix for issue556) and the improvements to how MathJax is configured and loaded. This is not finished yet, but at least the files for the localization, UI, and TeX parsing are now loaded via MediaWiki's resource loader.

Patch-set: 4
Reviewer: Frédéric Wang <1025@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0