Update patch set 6
Patch Set 6:

Thanks for the code review.
The method is really very similar... maybe I should expand my MathSearch project to code search as well... however there is a small difference.
ll 368.
$cgi .= wfArrayToCgi( $v, null, $key . "[$k]" );

Adds [] if $k is empty. So that won't work for me unless I replaced '[]' by ''. But I'm not really sure if you want me to do this to get this change merged.

Patch-set: 6