Update patch set 9
Patch Set 9:

With the last version of this patch, all the JS files should now be loaded with MediaWiki's resource loader. There are two small images that are not used in stylesheets, so they are loaded normally. The Web fonts are still loaded with MathJax's internal code.

Obviously, the big issue with this change is that it makes maintenance harder (redefinition of MathJax functions, tables files / modules). But perhaps that could be improved later with some scripts.

This is for the dev branch, since it relies on MathJax 2.3 which is not released yet.

Patch-set: 9
Label: Verified=0