Update patch set 1
Patch Set 1:

This implements the change discussed between MediaWiki and the MathJax team: it makes texvc and MathJax produce the same HTML output so that we could potentially expose the MathJax mode to anonymous users. Also, it modifies the wiki2jax preprocessor to use the texvc PNG images as a preview before MathJax rendering, as proposed by Peter.

Physikerwelt said he was against making the MathJax mode depends on texvc, but I'm not sure why. I understand that it adds loading of PNG images in addition to MathJax scripts so that's less optimal. But in my opinion, the LaTeXML mode (+Gecko native's MathML) is best to demonstrate how we could get a fast output with MathML while the MathJax mode is more intended to demonstrate how we can get a beautiful output with MathML. Having the image preview is way workaround the rendering delay due to MathJax's slowness.

Patch-set: 1
Label: Verified=0