Update patch set 1
Patch Set 1:

> I did not test that, but for me it reads as it forces everyone to use MathJax, except if you disable javascript in your browser.

No, it uses MathJax in the MW_MATH_MATHJAX mode and PNG in the MW_MATH_PNG as usual. The only difference is that the PNG images are used in the MW_MATH_MATHJAX as a preview.

> Furthermore I removes the mathjax mode that is used by several private wikis, that have neither texvc nor LaTeX installed on their webserver.

So what about keeping the MW_MATH_MATHJAX mode and adding a MW_MATH_PNG_MATHJAX mode instead? That will allow the MW_MATH_PNG_MATHJAX to be exposed to anonymous users.

Patch-set: 1