Update patch set 1
Patch Set 1:

That means you check the mode inside the matjax javascript... maybe a comment could be added to the php file?
I think in the mathjax mode it should be figured out automatically if latexml is installed. You could use the error messages from the texvc php file.

Patch-set: 1