Update patch set 2
Patch Set 2:

For LaTeXML, you've used the MathML "display" attribute, not the "displaystyle" attribute. I think that's what we want: only a parameter "display" that indicates whether equations should be in their own centered paragraph + rendered in display style (display=block) or just inside a paragraph of text (display=inline). So we don't want to bother with this "displaystyle" parameter but just have one "display" parameter. See my comment about how to pass it to texvc. Of course, since the output is an image you'll still need to create the appropriate HTML output to put that image in a centered paragraph when display=block.

As a consequence, we don't want to verify \displaystyle, which may be used in display=inline equation like $\displaystyle{\sum_a^b}$. LaTeXML, texvc or MathJax will automatically understand that command and do the appropriate rendering modification (e.g. adding displaystyle="true" to the MathML code).

guessDisplaytyleFromTex is just looking for a substring, which is very poor way to guess. For example $\text{equation} \in \text{mathematics}$ is display=inline. I think what we want is just to check the string via preg_match:

1) /^\w*\\begin/ => display=block (I think all the Math environments are display=block, but not 100% sure, please check)
2) /^\w*\$\$(.*)\$\$\w*/s => display=block
3) /^\w*\$(.*)\$\w*/s => display=inline
4) otherwise => display=inline

Of course, 2) should be executed before 3) and in both cases you want to use the returned subpattern (i.e. ignore the white spaces and dollars). In 4), if we want to preserve the current rendering, then we'll also want to concatenate "\displaystyle{" and "}" before sending it to texvc/latexml, just like you've done.

Patch-set: 2